code/ci-360-manage-contact-preferences/requirements.txt
PyJWT==2.10.1
PyJWT==2.10.1
Requests==2.32.5
tqdm==4.67.1
